requirements.txt
django==4.1.3                                  # web framework itself
djangorestframework==3.14.0                    # creating rest APIs
psycopg2-binary==2.9.5                         # postgres database adapter
requests>=2.31.0
django-cors-headers==3.13.0
django_db_logger==0.1.12
tzdata==2022.6
dj_database_url==1.0.0
django-environ==0.9.0                          # environment variables utility
pyyaml==6.0.1                                    # yml parser
whitenoise==6.2.0                              # enabling Django to serve its own static files
coverage==6.5.0                                # for test coverage
django-simple-history                         # To maintain history of table data
overrides                                     # Decorators to assert function overrides
openpyxl                                      # dependency for reading excel files through pandas
Twisted>=22.4                                  # daphne dependency package with specific version to fix CVE vulnerability
django-filter                                 # DRF package to add filtering on viewsets
drf-standardized-errors                       # DRF package to standerdize all 5xx and 4xx errors
channels==3.0.5                                # Prerequisite for django-eventstream
django-eventstream                            # For server sent events
daphne                                      # application server
django-helusers                               # Tunnistamo user inegration
social-auth-app-django
